Add ability to force user to change password
When injecting the user password through the ConfigDrive,
the password is stored in a non-encrypted format.

Because the ConfigDrive is readable by any non-administrator users,
this could cause serious security issues, one being the compromising
of the Administrator account by an unauthorized user.

This option adds the ability to force the user to change his password
after the next logon by expiring it, making the one stored in
the ConfigDrive obsolete and irrelevant from a security point of view.

This option is disabled by default for backward compatibility reason.